The bug solved itself? Or maybe it was adding .userDetailsService(applicationUserService) to the rememberMe builder chain in ApplicationSecurityConfig.configure? https://stackoverflow.com/questions/46421185/remember-me-not-working-throws-java-lang-illegalstateexception-userdetailsse https://docs.spring.io/spring-security/site/docs/current/reference/html5/#modules
Who knows..? It was running fine without that fix on lovelace, and now it seems to be running fine on pascaline when I remove the fix... I also did a clean build. Maybe that cleared out some gunk in the Spring framework pipes?

Either way. We done.

Feature works. We can now use a real DB to store authentication users if we want. woo